src/subagent/system_designer_beta/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
anthropic>=0.40.0
python-dotenv>=1.0.0
pyyaml>=6.0
openpyxl>=3.1.0
python-docx>=1.1.0
